+++ packages/nimble-components/src/anchor/tests/anchor.spec.ts
@@ -2,7 +2,10 @@ import { html } from '@microsoft/fast-element';
 import { Anchor, anchorTag } from '..';
 import { waitForUpdatesAsync } from '../../testing/async-helpers';
 import { fixture, Fixture } from '../../utilities/tests/fixture';
-import { getSpecTypeByNamedList } from '../../utilities/tests/parameterized';
+import {
+    getSpecTypeByNamedList,
+    parameterizeNamedList
+} from '../../utilities/tests/parameterized';
 
 async function setup(): Promise<Fixture<Anchor>> {
     return fixture<Anchor>(html`<nimble-anchor></nimble-anchor>`);
@@ -90,5 +93,73 @@ describe('Anchor', () => {
                 );
             });
         }
+    });
+
+    describe('contenteditable behavior', () => {
+        let innerAnchor: HTMLAnchorElement;
+
+        beforeEach(async () => {
+            await connect();
+            innerAnchor = element.shadowRoot!.querySelector('a')!;
+        });
+
+        it('has undefined property value and inner anchor isContentEditable is false by default', () => {
+            expect(element.contentEditable).toBeUndefined();
+            expect(innerAnchor.isContentEditable).toBeFalse();
+        });
+
+        const interestingValues = [
+            { name: '', expected: true, skipTag: '' },
+            { name: 'true', expected: true, skipTag: '' },
+            { name: 'false', expected: false, skipTag: '' },
+            { name: 'plaintext-only', expected: true, skipTag: '#SkipFirefox' },
+            { name: 'inherit', expected: false, skipTag: '' },
+            { name: 'badvalue', expected: false, skipTag: '' }
+        ] as const;
+
+        parameterizeNamedList(interestingValues, (spec, name, value) => {
+            spec(
+                `inner anchor isContentEditable is ${value.expected.toString()} when attribute set to "${name}" ${
+                    value.skipTag
+                }`,
+                async () => {
+                    element.setAttribute('contenteditable', name);
+                    await waitForUpdatesAsync();
+                    expect(innerAnchor.isContentEditable).toEqual(
+                        value.expected
+                    );
+                }
+            );
+        });
+
+        parameterizeNamedList(interestingValues, (spec, name, value) => {
+            spec(
+                `inner anchor isContentEditable is ${value.expected.toString()} when property set to "${name}" ${
+                    value.skipTag
+                }`,
+                async () => {
+                    element.contentEditable = name;
+                    await waitForUpdatesAsync();
+                    expect(innerAnchor.isContentEditable).toEqual(
+                        value.expected
+                    );
+                }
+            );
+        });
+    });
+
+    describe('with contenteditable without value', () => {
+        async function setupWithContenteditable(): Promise<Fixture<Anchor>> {
+            return fixture<Anchor>(
+                html`<nimble-anchor contenteditable></nimble-anchor>`
+            );
+        }
+
+        it('acts like value is "true"', async () => {
+            ({ element, connect, disconnect } = await setupWithContenteditable());
+            await connect();
+            const innerAnchor = element.shadowRoot!.querySelector('a')!;
+            expect(innerAnchor.isContentEditable).toBeTrue();
+        });
     });
 });